[Inductor] move alignment tests to a separate file
[ghstack-poisoned]
