+++ test/inductor/test_torchinductor.py
@@ -13248,99 +13248,6 @@ def forward(self, arg0_1: "Sym(s77)", arg1_1: "Sym(s27)", arg2_1: "Sym(s53)", ar
         self.common(fn, (1, x))
         self.common(fn, (2, x))
 
-    def test_unaligned_input(self):
-        def fn(x):
-            return torch.nn.functional.relu(x)
-
-        x = torch.randn(1024 + 16, device=self.device)[1:-15]
-        # TODO (malfet): Investigate failures on MacOS-14
-        with (
-            contextlib.nullcontext()
-            if self.device != "mps" or MACOS_VERSION >= 15.0
-            else self.assertRaises(AssertionError)
-        ):
-            self.common(fn, (x,), check_lowp=False)
-
-    def test_unaligned_input_2d(self):
-        def fn(x):
-            return torch.nn.functional.relu(x)
-
-        x = torch.randn(1024, 1024 + 16, device=self.device)[:, 1:-15]
-        self.common(fn, (x,), check_lowp=False)
-
-    def test_alignment_without_custom_op(self):
-        def fn(x):
-            a = torch.nn.functional.relu(x)
-            b = (3 * a)[1:-15]
-            c = torch.cos(b)
-            return c
-
-        x = torch.randn(1024 + 16, device=self.device)
-        self.common(fn, (x,), check_lowp=False)
-
-    @config.patch(implicit_fallbacks=True)
-    def test_no_align_for_custom_op(self):
-        def slice1d(x):
-            return (3 * x)[1:-15]
-
-        def slice1d_meta(x):
-            return torch.empty_like(x)[1:-15]
-
-        define_custom_op_for_test("slice1d", slice1d, slice1d_meta)
-
-        def fn(x):
-            a = torch.nn.functional.relu(x)
-            b = torch.ops.test.slice1d(a)
-            c = torch.cos(b)
-            return c
-
-        x = torch.randn(1024 + 16, device=self.device)
-        self.common(fn, (x,), check_lowp=False)
-
-    @config.patch(implicit_fallbacks=True)
-    def test_no_align_for_custom_op_2d(self):
-        def slice2d(x):
-            return (3 * x)[..., 1:-15]
-
-        def slice2d_meta(x):
-            return torch.empty_like(x)[..., 1:-15]
-
-        define_custom_op_for_test("slice2d", slice2d, slice2d_meta)
-
-        def fn(x):
-            a = torch.nn.functional.relu(x)
-            b = torch.ops.test.slice2d(a)
-            c = torch.cos(b)
-            return c
-
-        x = torch.randn(1024, 1024 + 16, device=self.device)
-        self.common(fn, (x,), check_lowp=False)
-
-    @config.patch(implicit_fallbacks=True, alignment_asserts=True)
-    @skip_if_cpp_wrapper(
-        "Inductor does not generate alignment assertion for cpp_wrapper right now"
-    )
-    def test_incorrect_meta_for_custom_op_2d(self):
-        def slice2d(x):
-            return (3 * x)[..., 1:-15]
-
-        def slice2d_meta(x):
-            return torch.empty_like(x)[..., 0:-16]
-
-        define_custom_op_for_test("slice2d_incorrect_meta", slice2d, slice2d_meta)
-
-        def fn(x):
-            a = torch.nn.functional.relu(x)
-            b = torch.ops.test.slice2d_incorrect_meta(a)
-            c = torch.cos(b)
-            return c
-
-        x = torch.randn(1024, 1024 + 16, device=self.device)
-
-        expected_error = "Expect the tensor to be 16 bytes aligned. Fail due to storage_offset=1 itemsize=4"
-        with self.assertRaisesRegex(AssertionError, expected_error):
-            self.common(fn, (x,), check_lowp=False)
-
 
 @dataclasses.dataclass
 class TestFailure:
